Separate jmt.hpp from trajectory.hpp
